Add casting API to the C++ side of the clocks module
This commit adds methods necessary to cast durations and type_points to and
from `std::chrono::nanoseconds` after passing them across ffi as `int64_t`.

C++ side of the timer module needs those methods. They are implemented in
`clocks.{hh, cc}` because this is a more general place for them (they might
be used in other place than the timer module).

Some definicions introduced in this commit enabled simplification of already
existent code in `timer.cc`. This is also done in this commit.