add seastar_test_guard module
Although seastar seems to support starting and stopping the runtime
multiple times in a sequence, running multiple runtimes in parallel in
the same process doesn't work. In order to support parallel unit tests,
a module is introduced that provides mutual exclusion for tests that use
the seastar runtime.

The `acquire_guard_for_seastar_test` function should be called by each
test which creates a seastar runtime, and the result held until the test
finishes. In the future when we get the seastar::test macro, this
function should be called transparently by the code generated by the
macro.