add cxx_async_local_future module
We are using cxx_async for passing futures/coroutines between Rust and
C++. The library was clearly made with work-stealing schedulers in mind,
so the Rust side imposes a `Send` requirement on the futures passed to
C++. However, seastar has a shared-nothing model so futures spawned on
tasks won't migrate between threads, therefore the `Send` requirement is
too much.

We are working around this limitation by lying to cxx_async - we tell it
that our futures are indeed `Send` even if they don't (and frequently
won't). It's technically speaking unsound because now the _safe_ code
needs to be careful not to accidentally pass a cxx_async wrapper future
to another thread - however, it's not super bad if we only use the
wrapper futures within the internals of seastar-rs. In the future, we
might improve the safety of the interface further, which would probably
require writing procedural macros that piggy-back on cxx_async's macros
but specify different bounds for the futures being wrapped.
Alternatively, we can investigate the possibility of dropping cxx_async
and writing a bespoke library that translates Rust futures to seastar
futures.

This commit introduces the cxx_async_local_future module which defines
an extension trait for futures generated by cxx_async and allows putting
non-sync futures into cxx_async futures.