Add Gate::try_enter
This commit adds `try_enter` method to the `Gate` type. This method
returns an instance of `GateHolder` that leaves the gate when
dropped. Such an implementation allows us to simulate a RAII
pattern used in `seastar::gate` and `seastar::gate::holder`.

There is also GateClosedError which is an equivalent of
seastar::gate::gate_closed_exception.