Add SchedulingGroup::{create, destroy, rename}
This commit introduces 3 methods/functions of `SchedulingGroup` that
correspond to friend functions of `seastar::scheguduling_group` class:
`create`, `destroy`, `rename`. We do this this way because there is such
concept as friend funcions in Rust. Besides, this approach seems to have
no disadvantages.

Implementing `create` is quite tricky because we can't await a future
that waits for `scheduling_group` (or `UniquePtr<scheduling_group>`)
in Rust. The workaround here is creating `UniquePtr<scheduling_group>`
in Rust, passing its reference to `create_sg` and changing it to new
`UniquePtr` with the value returned by `seastar::create_scheduling_group`.