Set up callback passing across ffi
This commit does everything that is needed to pass callbacks implementing
the `FnMut()` trait across ffi.

Apart from the callback we pass to C++:
- a caller – created by `get_caller` function, calls the callback,
- a dropper – created by `get_dropper` function, drops the callback.

To pass the callback to C++ we will put it inside a `Box` and cast it to
`*mut u8` − a substitue for `*mut c_void` that is not supported by `cxx`.

We want to drop the callback in a right moment. That's why we create
the `callback_object` class which is a callable object that drops the
callback when its destructor is called. We overwrite move constructor and
move operator of the `callback_object` to avoid droppping the callback
in a temporary instance's destructor. We will pass instances of this class
instead of pure callbacks to timers' `set_callback` methods.