Add the net_ffi module
This commit exposes the part of Seastar's network API necessary for this
example.

We don't expose `ffi::accept` because it does not work in a user friendly way.
This function changes `socket` by reference instead of returning it. That's
why a simpler wrapper is exposed.

The rest of the code is pretty straightforwad.